Fri, Apr 12, 2024, 9:05 AM +02:00

+++ collatz.c
@@ -7,7 +7,7 @@ int main(void)
 
 float n = get_int("n = ");
 int z = 0;
-float x = round(n)/2;
+float x = (float) roundf(n)/2;
 float y = n/2;
 
 while (n > 1)
